requirements.txt
flask==3.0.0
flask-socketio==5.3.5
python-engineio==4.3.4
python-socketio==5.8.0
python-docx==1.1.0